requirements.txt
pyautogen>=0.8.3
litellm>=1.64.1
python-dotenv>=1.0.1
phoenix>=0.9.1
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.6.4
pydantic-settings>=2.1.0
requests>=2.31.0
python-multipart>=0.0.6
arize>=7.0.0
arize-phoenix>=0.1.0
pytest>=8.3.5
pytest-asyncio>=0.23.5
pytest-timeout==2.3.1
langfuse>=2.0.0
# Optional dependencies
# mem0>=1.0.0  # Optional: Memory management
httpx>=0.25.2
aiohttp>=3.9.3
asyncio==3.4.3
ollama==0.1.6
numpy==1.26.4
pandas==2.2.1 